Extract useTheme hook from App component
